Enrichment tool updated fields for enrichemnt models
