--- src/app/common/iddescription.ts
@@ -1,7 +1,0 @@
-export class IdDescriptionName {
-  public constructor(
-    public readonly id: string,
-    public readonly description: string,
-    public readonly name: string
-  ) { }
-}
