+++ src/app/enrichment-models/iddescription.ts
@@ -1,0 +1,8 @@
+export class IdDescriptionName {
+  public constructor(
+      public readonly id: string,
+      public readonly description: string,
+      public readonly name: string
+  ) { }
+}
+
